Add Errors for FileSystem

+++ src/util/FileSystem.ts
@@ -1,4 +1,5 @@
 import { copyFileSync, existsSync, mkdirSync, readdirSync, rmdirSync, statSync, unlinkSync } from "fs";
+import { findSourceMap } from "module";
 import { FileSystemAdapter, Notice } from "obsidian";
 import { dirname, join } from "path";
 
@@ -11,39 +12,43 @@ export function getAllFiles(path: string[]): string[] {
 	let pathSections: string[] = [];
 	let files: string[] = [];
 
-	// Check path contains path placeholder
-	if (join(...path).includes('\\*\\')) {
-		pathSections = join(...path).split('\\*\\');
+	try {
+		// Check path contains path placeholder
+		if (join(...path).includes('\\*\\')) {
+			pathSections = join(...path).split('\\*\\');
 
-		if (pathSections.length > 0) {
-			// Get existing paths for placeholders
-			let pathContent = readdirSync(pathSections[0]);
+			if (pathSections.length > 0) {
+				// Get existing paths for placeholders
+				let pathContent = readdirSync(pathSections[0]);
 
-			// Add all combined files
-			pathContent.forEach(value => {
-				const joinedPath = join(pathSections[0], value, ...pathSections.filter((value, index) => index > 0));
-				files = files.concat(getAllFiles([joinedPath]));
+				// Add all combined files
+				pathContent.forEach(value => {
+					const joinedPath = join(pathSections[0], value, ...pathSections.filter((value, index) => index > 0));
+					files = files.concat(getAllFiles([joinedPath]));
+				});
+			}
+		}
+		// Check path contains file placeholder
+		else if (join(...path).endsWith('\\*')) {
+			pathSections = join(...path).split('\\*');
+
+			let pathContent = readdirSync(pathSections[0]).map(value => join(pathSections[0], value));
+			files = pathContent.filter((value) => {
+				return !statSync(value).isDirectory();
 			});
 		}
+		// Path doesn't exist
+		else if (!existsSync(join(...path))) {
+			return [];
+		}
+		// Path is file
+		else if (!statSync(join(...path)).isDirectory()) {
+			return [];
+		}
+		return files;
+	} catch (e) {
+		throw e;
 	}
-	// Check path contains file placeholder
-	else if (join(...path).endsWith('\\*')) {
-		pathSections = join(...path).split('\\*');
-
-		let pathContent = readdirSync(pathSections[0]).map(value => join(pathSections[0], value));
-		files = pathContent.filter((value) => {
-			return !statSync(value).isDirectory();
-		});
-	}
-	// Path doesn't exist
-	else if (!existsSync(join(...path))) {
-		return [];
-	}
-	// Path is file
-	else if (!statSync(join(...path)).isDirectory()) {
-		return [];
-	}
-	return files;
 }
 
 /**
@@ -55,33 +60,37 @@ export function getAllSubPaths(path: string[]): string[] {
 	let pathSections: string[] = [];
 	let paths: string[] = [];
 
-	// Check path contains placeholder
-	if (join(...path).includes('\\*\\')) {
-		pathSections = join(...path).split('\\*\\');
+	try {
+		// Check path contains placeholder
+		if (join(...path).includes('\\*\\')) {
+			pathSections = join(...path).split('\\*\\');
 
-		if (pathSections.length > 0) {
-			// Get existing paths for placeholder
-			let pathContent = readdirSync(pathSections[0]);
+			if (pathSections.length > 0) {
+				// Get existing paths for placeholder
+				let pathContent = readdirSync(pathSections[0]);
 
-			// Add all combined paths
-			pathContent.forEach(value => {
-				const joinedPath = join(pathSections[0], value, ...pathSections.filter((value, index) => index > 0));
-				paths = paths.concat(getAllSubPaths([joinedPath]));
+				// Add all combined paths
+				pathContent.forEach(value => {
+					const joinedPath = join(pathSections[0], value, ...pathSections.filter((value, index) => index > 0));
+					paths = paths.concat(getAllSubPaths([joinedPath]));
+				});
+			}
+		}
+		// Path doesn't exist
+		else if (!existsSync(join(...path))) {
+			return [];
+		}
+		// Get subpath in path
+		else {
+			let pathContent = readdirSync(join(...path)).map(value => join(...path, value));
+			paths = pathContent.filter((value) => {
+				return statSync(value).isDirectory();
 			});
 		}
+		return paths;
+	} catch (e) {
+		throw e;
 	}
-	// Path doesn't exist
-	else if (!existsSync(join(...path))) {
-		return [];
-	}
-	// Get subpath in path
-	else {
-		let pathContent = readdirSync(join(...path)).map(value => join(...path, value));
-		paths = pathContent.filter((value) => {
-			return statSync(value).isDirectory();
-		});
-	}
-	return paths;
 }
 
 /**
@@ -90,23 +99,23 @@ export function getAllSubPaths(path: string[]): string[] {
  * @param targetPath The target file
  */
 export function keepNewestFile(sourcePath: string[], targetPath: string[]) {
-	const sourceFile = join(...sourcePath);
-	const targetFile = join(...targetPath);
+	try {
+		const sourceFile = join(...sourcePath);
+		const targetFile = join(...targetPath);
 
-	// Keep newest file
-	if (existsSync(sourceFile) && (!existsSync(targetFile) || statSync(sourceFile).mtime > statSync(targetFile).mtime)) {
-		// Check target path exist
-		if (!ensurePathExist([dirname(targetFile)])) {
-			return;
+		// Keep newest file
+		if (existsSync(sourceFile) && (!existsSync(targetFile) || statSync(sourceFile).mtime > statSync(targetFile).mtime)) {
+			// Check target path exist
+			ensurePathExist([dirname(targetFile)])
+			copyFileSync(sourceFile, targetFile);
 		}
-		copyFileSync(sourceFile, targetFile);
-	}
-	else if (existsSync(targetFile)) {
-		// Check target path exist
-		if (!ensurePathExist([dirname(sourceFile)])) {
-			return;
+		else if (existsSync(targetFile)) {
+			// Check target path exist
+			ensurePathExist([dirname(sourceFile)])
+			copyFileSync(targetFile, sourceFile);
 		}
-		copyFileSync(targetFile, sourceFile);
+	} catch (e) {
+		throw e;
 	}
 }
 
@@ -116,21 +125,23 @@ export function keepNewestFile(sourcePath: string[], targetPath: string[]) {
  * @param targetPath The target file
  * @returns Copy was successfull
  */
-export function copyFile(sourcePath: string[], targetPath: string[]): boolean {
-	const sourceFile = join(...sourcePath);
-	const targetFile = join(...targetPath);
+export function copyFile(sourcePath: string[], targetPath: string[]) {
+	try {
+		const sourceFile = join(...sourcePath);
+		const targetFile = join(...targetPath);
 
-	// Check source exist
-	if (!existsSync(sourceFile)) {
-		return false;
-	}
-	// Check target path exist
-	if (!ensurePathExist(targetPath.slice(0, targetPath.length - 1))) {
-		return false;
-	}
+		// Check source exist
+		if (!isValidPath([sourceFile]) || !existsSync(sourceFile)) {
+			throw Error(`Source file does not exist! ${sourceFile}`);
+		}
+		// Check target path exist
+		isValidPath([...targetPath])
+		ensurePathExist(targetPath.slice(0, targetPath.length - 1));
 
-	copyFileSync(sourceFile, targetFile);
-	return true;
+		copyFileSync(sourceFile, targetFile);
+	} catch (e) {
+		throw e;
+	}
 }
 
 /**
@@ -139,36 +150,36 @@ export function copyFile(sourcePath: string[], targetPath: string[]): boolean {
  * @param targetPath The target path where to copy the subfolders/files to
  */
 export function copyFolderRecursiveSync(sourcePath: string[], targetPath: string[]) {
-	const source = join(...sourcePath);
-	const target = join(...targetPath);
+	try {
+		const source = join(...sourcePath);
+		const target = join(...targetPath);
 
-	// Check source is a valid path and exist
-	if (!isValidPath([source]) || !existsSync(source)) {
-		return false;
-	}
-	// Check target is a valid path and ensure exist 
-	if (!isValidPath([target]) || !ensurePathExist([target])) {
-		new Notice(`Failed to copy folder!`);
-		return false;
-	}
-
-	// Files in source
-	const files = readdirSync(source);
-
-	files.forEach(file => {
-		const sourceFile = join(source, file);
-		const targetFile = join(target, file);
-
-		if (statSync(sourceFile).isDirectory()) {
-			// Copy files in subpath
-			copyFolderRecursiveSync([sourceFile], [targetFile]);
-		} else {
-			// Copy file
-			copyFileSync(sourceFile, targetFile);
+		// Check source is a valid path and exist
+		if (!isValidPath([source]) || !existsSync(source)) {
+			throw Error('Source path does not exist!');
 		}
-	});
+		// Check target is a valid path and ensure exist 
+		isValidPath([target])
+		ensurePathExist([target])
 
-	return true;
+		// Files in source
+		const files = readdirSync(source);
+
+		files.forEach(file => {
+			const sourceFile = join(source, file);
+			const targetFile = join(target, file);
+
+			if (statSync(sourceFile).isDirectory()) {
+				// Copy files in subpath
+				copyFolderRecursiveSync([sourceFile], [targetFile]);
+			} else {
+				// Copy file
+				copyFileSync(sourceFile, targetFile);
+			}
+		});
+	} catch (e) {
+		throw e;
+	}
 }
 
 /**
@@ -177,12 +188,18 @@ export function copyFolderRecursiveSync(sourcePath: string[], targetPath: string
  * @param recursive [true] Indicates whether parent folders should be created.
  * @returns Returns ``true`` if the path exists, ``false`` if failed to create the path.
  */
-export function ensurePathExist(path: string[], recursive = true): boolean {
-	// If path not exist create it 
-	if (!existsSync(join(...path))) {
-		mkdirSync(join(...path), { recursive });
+export function ensurePathExist(path: string[], recursive = true) {
+	try {
+		// If path not exist create it 
+		if (!existsSync(join(...path))) {
+			mkdirSync(join(...path), { recursive });
+			if (!existsSync(join(...path))) {
+				throw Error('Could not create path!');
+			}
+		}
+	} catch (e) {
+		throw e;
 	}
-	return existsSync(join(...path));
 }
 
 /**
@@ -191,15 +208,12 @@ export function ensurePathExist(path: string[], recursive = true): boolean {
  * @returns True if is Valid
  */
 export function isValidPath(path: string[]) {
-	try {
-		// Check is not an empty string
-		if (join(...path) === "") {
-			return false;
-		}
-		// accessSync(path, constants.F_OK);
-	} catch (err) {
+	// Check is not an empty string
+	if (join(...path) === "") {
 		return false;
 	}
+	// accessSync(path, constants.F_OK);
+
 	return true;
 }
 
@@ -208,23 +222,27 @@ export function isValidPath(path: string[]) {
  * @param path The folder to remove
  */
 export function removeDirectoryRecursiveSync(path: string[]) {
-	const pathS = join(...path);
+	try {
+		const pathS = join(...path);
 
-	if (existsSync(pathS)) {
-		readdirSync(pathS).forEach(file => {
-			const filePath = join(pathS, file);
+		if (existsSync(pathS)) {
+			readdirSync(pathS).forEach(file => {
+				const filePath = join(pathS, file);
 
-			if (statSync(filePath).isDirectory()) {
-				// Recursively remove subdirectories
-				removeDirectoryRecursiveSync([filePath]);
-			} else {
-				// Remove files
-				unlinkSync(filePath);
-			}
-		});
+				if (statSync(filePath).isDirectory()) {
+					// Recursively remove subdirectories
+					removeDirectoryRecursiveSync([filePath]);
+				} else {
+					// Remove files
+					unlinkSync(filePath);
+				}
+			});
 
-		// Remove the empty directory
-		rmdirSync(pathS);
+			// Remove the empty directory
+			rmdirSync(pathS);
+		}
+	} catch (e) {
+		throw e;
 	}
 }
 
@@ -237,5 +255,6 @@ export function getVaultPath() {
 	if (adapter instanceof FileSystemAdapter) {
 		return adapter.getBasePath();
 	}
-	return "";
+
+	return '';
 }
